+++ test_app/app/src/main/java/com/example/test_app/MainActivity.kt
@@ -1,11 +1,11 @@
 package com.example.test_app
 
 import android.content.DialogInterface
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
 
 class MainActivity : AppCompatActivity() {