+++ test_app/multi-modules/fullyIgnoredModule/src/androidTest/java/com/example/fullyIgnoredModule/ExampleInstrumentedTest.kt
@@ -1,0 +1,24 @@
+package com.example.fullyIgnoredModule
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.assertEquals
+import org.junit.Ignore
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Ignore
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("com.example.fullyIgnoredModule.test", appContext.packageName)
+    }
+}