+++ test_app/multi-modules/testModuleCommon/src/main/java/com/example/common/CommonInstrumentedTest.kt
@@ -1,0 +1,15 @@
+package com.example.common
+
+import org.junit.Assert
+
+fun testSampleMethod() {
+    Assert.assertEquals(1, "1".toInt())
+}
+
+fun anotherSampleTestMethod() {
+    Assert.assertEquals(2, "2".toInt())
+}
+
+fun ignoredTest() {
+    Assert.assertEquals(1, 1)
+}